+++ quill/include/quill/sinks/Sink.h
@@ -52,15 +52,19 @@ public:
    * @param log_timestamp Timestamp of the log event.
    * @param thread_id ID of the thread.
    * @param thread_name Name of the thread.
+   * @param process_id Process Id
    * @param logger_name Name of the logger.
    * @param log_level Log level of the message.
    * @param named_args Vector of key-value pairs of named args
    * @param log_message The log message.
+   * @param log_statement The log statement.
    */
-  QUILL_ATTRIBUTE_HOT virtual void write_log_message(
-    MacroMetadata const* log_metadata, uint64_t log_timestamp, std::string_view thread_id,
-    std::string_view thread_name, std::string_view logger_name, LogLevel log_level,
-    std::vector<std::pair<std::string, std::string>> const* named_args, std::string_view log_message) = 0;
+  QUILL_ATTRIBUTE_HOT virtual void write_log(MacroMetadata const* log_metadata, uint64_t log_timestamp,
+                                             std::string_view thread_id, std::string_view thread_name,
+                                             std::string const& process_id,
+                                             std::string_view logger_name, LogLevel log_level,
+                                             std::vector<std::pair<std::string, std::string>> const* named_args,
+                                             std::string_view log_message, std::string_view log_statement) = 0;
 
   /**
    * @brief Flushes the sink, synchronizing the associated sink with its controlled output sequence.
@@ -130,13 +134,13 @@ public:
    * @param thread_name Name of the thread.
    * @param logger_name Name of the logger.
    * @param log_level Log level of the message.
-   * @param log_message The log message.
+   * @param log_statement The log message.
    * @return True if the log record passes all filters, false otherwise.
    */
   QUILL_NODISCARD bool apply_all_filters(MacroMetadata const* log_metadata, uint64_t log_timestamp,
                                          std::string_view thread_id, std::string_view thread_name,
                                          std::string_view logger_name, LogLevel log_level,
-                                         std::string_view log_message)
+                                         std::string_view log_statement)
   {
     if (log_level < _log_level.load(std::memory_order_relaxed))
     {
@@ -167,10 +171,10 @@ public:
     {
       return std::all_of(_local_filters.begin(), _local_filters.end(),
                          [log_metadata, log_timestamp, thread_id, thread_name, logger_name,
-                          log_level, log_message](Filter* filter_elem)
+                          log_level, log_statement](Filter* filter_elem)
                          {
-                           return filter_elem->filter(log_metadata, log_timestamp, thread_id,
-                                                      thread_name, logger_name, log_level, log_message);
+                           return filter_elem->filter(log_metadata, log_timestamp, thread_id, thread_name,
+                                                      logger_name, log_level, log_statement);
                          });
     }
   }