+++ quill/include/quill/Backend.h
@@ -21,7 +21,7 @@ namespace quill
 {
 /** Version Info **/
 constexpr uint32_t VersionMajor{4};
-constexpr uint32_t VersionMinor{5};
+constexpr uint32_t VersionMinor{6};
 constexpr uint32_t VersionPatch{0};
 constexpr uint32_t Version{VersionMajor * 10000 + VersionMinor * 100 + VersionPatch};